GenAI-Chatbot/requirements.txt
streamlit
pypdf2
sentence-transformers
openai
langchain_community
langchain
faiss-cpu
tiktoken